pyproject.toml
[project]
name = "mediguide"
version = "1.0.0"
description = "Medical Advisory Platform - Backend API"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "sqlalchemy==2.0.23",
    "python-multipart==0.0.6",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "python-dotenv==1.0.0",
    "aiofiles==23.2.1",
    "google-auth==2.23.4",
    "google-auth-oauthlib==1.1.0",
    "google-auth-httplib2==0.1.1",
    "requests==2.31.0",
    "openai==1.3.5",
    "google-generativeai==0.3.1",
    "googlemaps==4.10.0",
    "Pillow==10.1.0",
    "pydantic==2.5.0",
    "pydantic-settings==2.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
